+++ poc-multiple-nav-hosts/src/main/java/com/adamkobus/compose/navigation/poc/multinavhost/green/ui/dialog/GreenDialogInteractions.kt
@@ -1,0 +1,5 @@
+package com.adamkobus.compose.navigation.poc.multinavhost.green.ui.dialog
+
+data class GreenDialogInteractions(
+    val onDismissClicked: () -> Unit
+)