+++ poc-multiple-nav-hosts/src/main/java/com/adamkobus/compose/navigation/poc/multinavhost/green/ui/home/GreenHomeScreenInteractions.kt
@@ -1,0 +1,5 @@
+package com.adamkobus.compose.navigation.poc.multinavhost.green.ui.home
+
+data class GreenHomeScreenInteractions(
+    val onNextClicked: () -> Unit
+)