+++ poc-multiple-nav-hosts/src/main/java/com/adamkobus/compose/navigation/poc/multinavhost/green/ui/next/GreenNextScreenInteractions.kt
@@ -1,0 +1,5 @@
+package com.adamkobus.compose.navigation.poc.multinavhost.green.ui.next
+
+data class GreenNextScreenInteractions(
+    val onOpenDialogClicked: () -> Unit
+)