+++ poc-multiple-nav-hosts/src/main/java/com/adamkobus/compose/navigation/poc/multinavhost/main/ui/home/HomeScreenInteractions.kt
@@ -1,0 +1,5 @@
+package com.adamkobus.compose.navigation.poc.multinavhost.main.ui.home
+
+data class HomeScreenInteractions(
+    val onContinueClicked: () -> Unit
+)