+++ poc-multiple-nav-hosts/src/main/java/com/adamkobus/compose/navigation/poc/multinavhost/yellow/ui/dialog/YellowDialogInteractions.kt
@@ -1,0 +1,5 @@
+package com.adamkobus.compose.navigation.poc.multinavhost.yellow.ui.dialog
+
+data class YellowDialogInteractions(
+    val onDismissClicked: () -> Unit
+)