+++ tutorial/src/main/java/com/adamkobus/compose/navigation/tutorial/TutorialApplication.kt
@@ -1,6 +1,7 @@
 package com.adamkobus.compose.navigation.tutorial
 
 import android.app.Application
+import android.util.Log
 import com.adamkobus.compose.navigation.ComposeNavigation
 import com.adamkobus.compose.navigation.tutorial.nav.OpenRandomItemIntentResolver
 import com.adamkobus.compose.navigation.tutorial.nav.TutorialNavActionVerifier
@@ -11,6 +12,7 @@ class TutorialApplication : Application() {
     override fun onCreate() {
         super.onCreate()
         ComposeNavigation
+            .setLogLevel(Log.VERBOSE)
             .addNavActionVerifiers(TutorialNavActionVerifier)
             .addNavIntentResolvers(OpenRandomItemIntentResolver)
     }